엔티티 설정 변경
