+++ src/user/dto/add-user.args.ts
@@ -1,0 +1,19 @@
+import { ArgsType, Field } from '@nestjs/graphql';
+import { IsEmail, IsNotEmpty, IsString } from 'class-validator';
+
+@ArgsType()
+export class AddUserArgs {
+  @IsEmail()
+  @Field()
+  email: string;
+
+  @IsNotEmpty()
+  @IsString()
+  @Field()
+  name: string;
+
+  @IsNotEmpty()
+  @IsString()
+  @Field()
+  password: string;
+}
