+++ src/user/user.resolver.ts
@@ -1,6 +1,7 @@
-import { Query, Resolver } from '@nestjs/graphql';
+import { Mutation, Query, Resolver } from '@nestjs/graphql';
 
-import { User } from '../entity/user';
+import { User } from '../entity/user.entity';
+import { AddUserArgs } from './dto/add-user.args';
 import { UserService } from './user.service';
 
 @Resolver()
@@ -10,5 +11,10 @@ export class UserResolver {
   @Query(() => [User], { nullable: 'items' })
   async users() {
     return this.userService.getUsers();
+  }
+
+  @Mutation(() => Boolean)
+  async addUser(args: AddUserArgs) {
+    return this.userService.addUser(args);
   }
 }
