+++ src/user/user.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@nestjs/common';
 
+import { User } from '../entity/user.entity';
 import { UserRepository } from '../repository/user.repository';
 
 @Injectable()
@@ -7,5 +8,9 @@ export class UserService {
   constructor(private readonly userRepository: UserRepository) {}
   async getUsers() {
     return this.userRepository.getUsers();
+  }
+
+  async addUser(args: Pick<User, 'email' | 'name' | 'password'>) {
+    return this.userRepository.addUser(args);
   }
 }
